Add QuizRunning functionality and enhance ShellUtils
The update encompasses the creation of QuizResults and QuizRunner classes to implement the Quiz running feature. It enhances the text coloring functionality in ShellUtils by adding style options. Furthermore, a new method is added to QuizService to get a Quiz object via its ID. ShellUtilsTest class was added to perform unit testing of the ShellUtils module. Last but not least, the QuizCommands class has been updated to include the run command, which utilizes the QuizRunner class to run quizzes.