Refactor DatesAndTimes class and add StringSplitting class
In the DatesAndTimes class, the method displayDurationAsString() has been refactored to return a formatted string instead of printing it directly. Removed the printing of duration units. In addition, a new class for string splitting named StringSplitting has been added which takes a string as input and splits it over specified width.